### 2017-11-06 3d6fc9a
docs for new Material release
