+++ docs/mkdocs.yml
@@ -2,12 +2,13 @@ site_name:           DynamicalBilliards.jl
 repo_url:            https://github.com/JuliaDynamics/DynamicalBilliards.jl
 site_description:    An easy-to-use, modular and extendable Julia package for Dynamical Billiard systems in two dimensions.
 site_author:         George Datseris
-theme:               'material'
 
-extra:
+theme:
+  name: 'material'
   palette:
     primary: 'deep purple'
     accent: 'cyan'
+extra:
   social:
     - type: 'github'
       link: 'https://github.com/JuliaDynamics/DynamicalBilliards.jl'
