### 2017-11-06 9bb0773
add email to appveyor
