+++ appveyor.yml
@@ -10,6 +10,8 @@ branches:
 
 notifications:
   - provider: Email
+    to:
+      - datseris.george@gmail.com
     on_build_success: false
     on_build_failure: true
     on_build_status_changed: false
