### 2017-11-14 #442 @Datseris
Hello, thanks for finally getting into this, I appreciate it.

The following seem to have relation to the Physics area I am interested in:

mlxd
jarvist
ahwillia

but still the connection is not that big... I hope it will be okay.

P.S.: Could you tell me how I can also apply to become a reviewer? I searched the website but I could not find where the application is. Do I just put my name in the google spreadsheet?
