### 2017-11-15 #442 @kyleniemeyer
Hi @Datseris, I can take care of that—aside from the existing JOSS reviewers, is there anyone closer to your domain who you might recommend? (It's ok if not)

As for adding your name to the reviewer list, you can use this link: http://joss.theoj.org/reviewer-signup.html
