### 2017-11-15 #442 @ChrisRackauckas
@ahwillia would be a good fit.
