### 2017-11-15 #442 @kyleniemeyer
@ChrisRackauckas thanks!

Hello @ahwillia @mlxd @jarvist @ahwillia, would you be willing to review this submission for JOSS? (I prefer to have at least two reviewers, but more are also welcome if you are interested!) Thanks!
