### 2017-11-15 #442 @Datseris
@kyleniemeyer unfortunately all the people that I can recommend have discussed with me the code in some way or another, therefore the conflict of interests would not go away...
