### 2017-11-15 #442 @kyleniemeyer
@Datseris that's all right, let's see if we can find someone else first. Having a discussion would not be a COI, just actually contributing to the code.
