### 2017-11-15 #442 @mlxd
Also happy to review this.
