### 2017-11-15 #458 @whedon
[REVIEW]: DynamicalBilliards.jl : An easy-to-use, modular and extendable Julia package for Dynamical Billiard systems in two dimensions.
**Submitting author:** @Datseris (<a href="http://orcid.org/0000-0002-6427-2385">George Datserus</a>)
**Repository:** <a href="https://github.com/JuliaDynamics/DynamicalBilliards.jl" target ="_blank">https://github.com/JuliaDynamics/DynamicalBilliards.jl</a>
**Version:** v1.6.1
