### 2017-11-15 bf2da55
Create ISSUE_TEMPLATE.md
