+++ ISSUE_TEMPLATE.md
@@ -1,0 +1,7 @@
+# Issue Description
+
+# Expected Behavior
+
+# MWE
+MWE stands for Minimal Working Example. 
+Please include a piece of code that can replicate your problem in a simple example.
