### 2017-11-15 #458 @ahwillia
Overall this looks excellent - really appreciate your attention to detail while putting this together!

@Datseris - what do you think about having an `examples/` folder containing jupyter notebooks? These could just contain the code that you have under `Tutorials` in the docs (but without the long text). That way people can flip back and forth between reading your docs and trying it. Right now there is an awkward copy-paste step.

Other than that I'm inclined to approve this. But we can wait for @mlxd to weigh in.
