### 2017-11-15 6907b8d
DOIS in paper.md
