### 2017-11-15 e782c37
add Contributing and Support sections in docs
