+++ docs/mkdocs.yml
@@ -42,5 +42,5 @@ pages:
   - Creating your own Obstacle: tutorials/own_obstacle.md
   - Examples: tutorials/examples.md
 - Physics: physics.md
-- Lyapunov Exponents: lyapunovs.md #commented out for now (until updated)
+- Lyapunov Exponents: lyapunovs.md
 - Library: basic/library.md
