### 2017-11-15 #458 @ahwillia
Yep - I'm fine with no jupyter notebooks, just a suggestion.

On Wed, Nov 15, 2017 at 3:21 PM, George Datseris <notifications@github.com>
wrote:

> Thanks for the kind words @ahwillia <https://github.com/ahwillia> . I've
> just added guidelines for issue reports.
>
> I will now do your other requests:
>
>    - Add DOIS to paper.md
>    - Add a dedicated paragraph on the homepage of the documentation about
>    support (we have our own chatroom on gitter)
>    - Add a dedicated contributors_guide file either in the repo first
>    page or in the documentation page.
>
> About the Jupyter notebooks... Is it possible to avoid this step? My only
> issue is that I have never used Jupyter and it make take a while/be tricky
> to set it up etc.
>
> I know it is kind of annoying to copy/paste that much but fortunately the
> documentation allows you to copy an entire code-block by clicking the top -
> right corner of each code block.
>
