### 2017-11-18 #458 @Datseris
This was simply a "dummy" name (indicated by the ***)... I expected users to put as a string wherever they want to save their files.

I will just replace `pwd()` since you are right, it is better.
