### 2017-11-18 629343a
add tempdir to docs instead of ***
