### 2017-11-18 #458 @Datseris
@kyleniemeyer the change is already done: https://juliadynamics.github.io/DynamicalBilliards.jl/latest/#julia-billiard-animation
