### 2017-11-18 #458 @kyleniemeyer
@Datseris ok, there are a few changes required for the paper.

1. The references need to be extracted from the Markdown paper file, put in a BibTeX .bib file, and then cited in the paper. You can see an example in the JOSS [author guidelines](http://joss.theoj.org/about#author_guidelines). This is necessary for our LaTeX-based compiling of the paper PDF, and also depositing the citation info in CrossRef.

2. The animated GIF you have is nice for the README, but it won't work for the paper, which is generated as a static PDF. Instead of the logo, could you perhaps include an example figure using results produced by the package (perhaps from one of the examples/tutorials), along with a brief description? That would be more useful and informative.

3. As for the actual content of the paper, the layout you have is not quite appropriate—again, it works for a README-type document, but not the paper. Links to your software repository will already be included in the paper (see [an example](https://github.com/openjournals/joss-papers/blob/joss.00460/10.21105.joss.00460.pdf) compiled PDF for a different paper under review), so please remove those from the top.

Also, to the summary or introductory paragraph, could you describe the research application of the software for a diverse, non-specialist audience? Many people may not be familiar with billiard systems in the context of research.

4. A few minor typos/suggested edits:
- "In [2] this is described perfectly, where Bunimovich and Vela-Arevalo (two pioneers in the field) summarize new insights in the field of dynamical billiards" -> "Bunimovich and Vela-Arevalo (two pioneers in the field) summarized new insights in the field of dynamical billiards [2]."
- "Our package has plenty of applications" is quite informal and not descriptive
- "which results to accuracies" -> "which results in accuracies"
- I would eliminate the use of italics to emphasize words throughout; it makes the paper read as more persuasive than is intended/appropriate, I think.
