### 2017-11-18 4bed02d
changes for editor Kyleniemeyer
