+++ paper/paper.md
@@ -17,69 +17,53 @@ date: 26 October 2017
 bibliography: none
 ---
 
-![Logo of DynamicalBilliards.jl](DynBil_logo---Animated.gif)
-
 # Summary
 
 [DynamicalBilliards.jl](http://orcid.org/0000-0002-6427-2385) is a package about
 two-dimensional (dynamical) billiard systems written in its entirety in Julia. It is easy to use, and easy to
 be extended. It is accompanied by a detailed documentation archive with a lot of tutorials, example code as well as information about the physical algorithms at play.
 
-Documentation:
-
-* [stable](https://juliadynamics.github.io/DynamicalBilliards.jl/stable/): https://juliadynamics.github.io/DynamicalBilliards.jl/stable/
-* [latest](https://juliadynamics.github.io/DynamicalBilliards.jl/latest/): https://juliadynamics.github.io/DynamicalBilliards.jl/latest/
-
-Repository page:
-
-* [GitHub](https://github.com/JuliaDynamics/DynamicalBilliards.jl): https://github.com/JuliaDynamics/DynamicalBilliards.jl
+The package is mainly used to simulate any kind of 2 dimensional system where particle motion is interrupted by collisions, like e.g. a gas infused with big molecules.
 
 # Features
 The features of the DynamicalSystems.jl, as of version v1.6.1, are:
 
 * Modular creation of a billiard table by putting together well-defined obstacles.
-* Creation of random initial conditions in *any* user-created billiard table.
+* Creation of random initial conditions in any user-created billiard table.
 * Propagation of point particles in such billiard tables.
 * Propagation of point particles in such billiard tables in magnetic fields.
 * Exact calculation of the collision times, collision positions, velocities etc..
 * Calculation of escape times of particles.
-* Rich implementation of Ray-splitting billiards: A particle may propagate *through* an obstacle, given some arbitrary transmission probability.
+* Rich implementation of Ray-splitting billiards: A particle may propagate through an obstacle, given some arbitrary transmission probability.
 * Calculation of Lyapunov exponents of trajectories (currently available only for propagation without magnetic fields).
 * Flexible speed; users can trade between accuracy and speed.
 * Extensive visualization library for plotting and animating particles and billiards.
 
 
-
 # Description
 Billiard systems have been used extensively in scientific research and played a
 crucial role in the development of Chaos theory. A famous example is the Sinai billiard
-(included in our package) which was one of the first low-dimensional systems proven to be ergodic [1].
+(easily simulated with our package) which was one of the first low-dimensional systems proven to be ergodic [@Sinai1970].
 
 Even though the study of billiard systems started decades ago, there are still new
-surprises and plenty of research to be done with them. In [2] this is described perfectly, where Bunimovich and Vela-Arevalo (two pioneers in the field) summarize
-new insights in the field of dynamical billiards.
+surprises and plenty of research to be done with them. Bunimovich and Vela-Arevalo (two pioneers in the field) summarized new insights in the field of dynamical billiards [@Bunimovich2015].
 
-Our package has plenty of applications, due to the large amount of features. It even offers the first time possibility
-of implementing ray-splitting billiards, since, to the best of our knowledge, there is no other open sourced project that offers this. In addition,
-the package is written in pure Julia [3] a new programming language with many advantages.
-Because Julia is *dynamic*, it allows users to experiment with billiards flexibly and *interactively* and have access to a very easy-to-use high level interface.
+Our package has plenty of applications, due to the large amount of features. In [@Datseris] we have used to approximate the movement of electrons in graphene, a two-dimensional material. Another application would be the motion of a particle in a gas composed of molecules much bigger than the particle, such that the collisions leave the molecules stationary. In addition, DynamicalBilliards.jl offers the first time possibility
[…]
