+++ paper/paper.bib
@@ -1,0 +1,84 @@
+@article{Sinai1970,
+abstract = {In this paper we consider dynamical systems resulting from the motion of a material point in domains with strictly convex boundary, that is, such that the operator of the second quadratic form is negative-definite at each point of the boundary, where the boundary is taken to be equipped with the field of inward normals. We prove that such systems are ergodic and are K -systems. The basic method of investigation is the construction of transversal foliations for such systems and the study of their properties.},
+author = {Sinai, Yakov G},
+doi = {10.1070/RM1970v025n02ABEH003794},
+file = {:C$\backslash$:/Users/datseris/AppData/Local/Mendeley Ltd./Mendeley Desktop/Downloaded/Sinai - 1970 - Dynamical systems with elastic reflections.pdf:pdf},
+issn = {0036-0279},
+journal = {Russian Mathematical Surveys},
+mendeley-groups = {Chaos {\&} Nonlinear Dynamics,Ragnaropaper Refs},
+number = {2},
+pages = {137},
+title = {{Dynamical systems with elastic reflections}},
+url = {http://stacks.iop.org/0036-0279/25/i=2/a=R05},
+volume = {25},
+year = {1970}
+}
+
+@article{Bunimovich2015,
+abstract = {“Chaos is found in greatest abundance wherever order is being sought.$\backslash$r$\backslash$nIt always defeats order, because it is better organized”$\backslash$r$\backslash$nTerry Pratchett$\backslash$r$\backslash$nA brief review is presented of some recent findings in the theory of chaotic dynamics. We also prove a statement that could be naturally considered as a dual one to the Poincar{\'{e}} theorem on recurrences. Numerical results demonstrate that some parts of the phase space of chaotic systems are more likely to be visited earlier than other parts. A new class of chaotic focusing billiards is discussed that clearly violates the main condition considered to be necessary for chaos in focusing billiards.},
+author = {Bunimovich, Leonid A. and Vela-Arevalo, Luz V.},
+doi = {10.1063/1.4916330},
+file = {:C$\backslash$:/Users/datseris/AppData/Local/Mendeley Ltd./Mendeley Desktop/Downloaded/Bunimovich, Vela-Arevalo - 2015 - Some new surprises in chaos (billiards).pdf:pdf},
+issn = {10541500},
+journal = {Chaos},
+mendeley-groups = {Chaos {\&} Nonlinear Dynamics},
+number = {9},
+title = {{Some new surprises in chaos (billiards)}},
+volume = {25},
+year = {2015}
+}
+
+@article{Julia,
+author = {Jeff Bezanson and Alan Edelman and Stefan Karpinski and Viral B. Shah},
+title = {Julia: A Fresh Approach to Numerical Computing},
+journal = {SIAM Review},
+volume = {59},
+number = {1},
+pages = {65-98},
+year = {2017},
+doi = {10.1137/141000671},
+
+URL = {
+        https://doi.org/10.1137/141000671
+
+},
+eprint = {
+        https://doi.org/10.1137/141000671
+
+}
+
+}
+
+@article{Solanpaa2016,
[…]
