### 2017-11-19 #458 @kyleniemeyer
@Datseris sure, that's fine—perhaps you could link on "detailed documentation archive" in the summary at the top of the paper?
