### 2017-11-19 #458 @Datseris
@kyleniemeyer I am glad we could come to an agreement. I have added an in-text hyperlink.

Is that all from my side?
