### 2017-11-19 #458 @arfon
Yeah, I'm debugging now. Getting errors reported back in here is somewhat tricky.
