### 2017-11-19 #458 @arfon
@Datseris - looks like you need to link to the `paper.bib` file. You currently have this set as `bibliography: none`
