### 2017-11-19 #458 @Datseris
@arfon I have changed it to `bibliography: paper.bib`. I hope this is enough.
