### 2017-11-19 #458 @Datseris
@kyleniemeyer I actually do not know what you mean with:
>The last step from you is archiving the final version of the software etc. (like using Zenodo) and then telling me the DOI here;

Isn't the DOI the duty of the Journal and not the author? I also do not know what "Zenodo" is.
