### 2017-11-19 #458 @kyleniemeyer
@Datseris we provide a DOI associated with the paper, but the author is responsible for archiving the accepted version of the software (i.e., your repository on GitHub). We do not archive the software and associated material, just the paper. You can see an example in a [recently published paper](http://www.theoj.org/joss-papers/joss.00451/10.21105.joss.00451.pdf) with the "Archive" link to the left.

This is from the JOSS About page:
> Upon successful completion of the review, deposit a copy of your (updated) software repository with a data-archiving service such as Zenodo or figshare, issue a DOI for the software archive, and update the review issue thread with your DOI.

I recommend using [Zenodo](https://zenodo.org), but [Figshare](https://figshare.com) is another option. Both are free, and used commonly for this purpose.
