### 2017-11-19 #458 @kyleniemeyer
Let's wait to see if any other changes are needed for the paper first, since it looks there this is still some problem.

@arfon it looks like whedon is raising the same error, but I can see in the `paper.md` that the bibliography seems set correctly now.
