### 2017-11-19 #458 @kyleniemeyer
@Datseris the link is actually present if you download the PDF, but the text is just not colored blue.

@arfon is there perhaps a `hyperref` setting for `href` links that is needed in the LaTeX template?
