### 2017-11-19 #458 @kyleniemeyer
(Sorry for the paper gen hiccups @Datseris, we are still testing out our new automated paper compilation)
