### 2017-11-19 #458 @Datseris
No problem! I also apologize because I did not realise that there was a link there. If you can add the standard hyperref color that would be much appreciated. If in the end of the day it appears not to be possible, then I can live with it being a "hidden link".

As a side note: how soon after the pdf generation works can I get a working DOI/reference to your journal? I am going to submit a publication (to another journal) next week and I am thinking if I can wait so I can include an official JOSS reference instead of a reference to the documentation page in my paper.
