### 2017-11-19 #458 @kyleniemeyer
@Datseris you will have the JOSS DOI basically as soon as we sort out this paper issue, so you should be able to include the full published JOSS reference.
