### 2017-11-19 #458 @Datseris
That is excellent. Since the paper seems to be in a final form besides this small hyperref detail, should I go ahead and perform this "Zenodo" process you described?
