### 2017-11-19 #458 @kyleniemeyer
Actually, I caught one final typo fix to make first:
change
> The package is mainly used to simulate any kind of 2 dimensional system where particle motion is interrupted by collisions, like e.g. a gas infused with big molecules.

to

>The package is mainly used to simulate any kind of two-dimensional system where particle motion is interrupted by collisions, e.g., a gas infused with big molecules.

Once you make that fix, please go ahead and archive the repository, and then report the DOI here. On Zenodo, you can do this manually by uploading a .zip (e.g.) of the repository; alternatively, Zenodo has a [nice GitHub integration](https://guides.github.com/activities/citable-code/) where it will automatically grab everything if you make a release on GitHub. (I do this for most of my own software packages.)
