### 2017-11-20 #458 @arfon
@kyleniemeyer @Datseris - the links in the paper are looking a bit better now.
