### 2017-11-20 #458 @kyleniemeyer
Yes, looks great! Ready to accept/publish.
