### 2017-11-20 #458 @arfon
I think you should be able to quote their usernames in back-ticks e.g. ````@arfon````
