### 2017-11-21 #458 @kyleniemeyer
@Datseris you can try using https://www.doi2bib.org/; we used to rely on it, but it isn’t always working.
