### 2017-11-21 #458 @danielskatz
also see https://github.com/openjournals/joss/issues/341
